🧹 refactor: extract confidence scoring into dedicated module
- Extract confidence scoring logic from `Scout._score_confidence` into `ConfidenceScorer` in `domain_scout/matching/scoring.py`.
- Move `_DomainAccum` to `domain_scout/models.py` and rename to `DomainAccumulator`.
- Move time normalization helpers to `domain_scout/models.py`.
- Update `Scout` and tests to use new components.
