+++ domain_scout/models.py
@@ -3,8 +3,12 @@
 from __future__ import annotations
 
 from datetime import datetime  # noqa: TC003 — Pydantic needs runtime import
+from typing import TYPE_CHECKING
 
 from pydantic import BaseModel, Field
+
+if TYPE_CHECKING:
+    from typing import Any
 
 
 class EntityInput(BaseModel):
@@ -116,3 +120,80 @@ class DeltaReport(BaseModel):
     warnings: list[DeltaWarning] = Field(default_factory=list)
     baseline_metadata: RunMetadata
     current_metadata: RunMetadata
+
+
+# --- Internal processing helpers and classes ---
+
+
+def _normalize_time(val: object) -> str | None:
+    """Normalize a datetime or string to ISO string for consistent comparison.
+
+    CT Postgres returns datetime objects, JSON API and cache return strings.
+    Normalizing to ISO strings prevents TypeError on mixed-type comparison.
+    """
+    if val is None:
+        return None
+    if isinstance(val, datetime):
+        return val.isoformat()
+    if isinstance(val, str):
+        if not val:
+            return None
+        try:
+            return datetime.fromisoformat(val).isoformat()
+        except ValueError:
+            return val
+    return str(val)
+
+
+def _parse_time(val: str | None) -> datetime | None:
+    """Parse an ISO 8601 string back to datetime for Pydantic output."""
+    if val is None:
+        return None
+    return datetime.fromisoformat(val)
+
+
+class DomainAccumulator:
+    """Mutable accumulator for evidence about a single domain."""
+
+    __slots__ = (
+        "sources",
+        "evidence",
+        "cert_org_names",
+        "first_seen",
+        "last_seen",
+        "resolves",
+        "rdap_org",
+        "confidence",
+    )
+
+    def __init__(self) -> None:
+        self.sources: set[str] = set()
+        self.evidence: list[EvidenceRecord] = []
+        self.cert_org_names: set[str] = set()
+        self.first_seen: str | None = None
+        self.last_seen: str | None = None
+        self.resolves: bool = False
+        self.rdap_org: str | None = None
+        self.confidence: float = 0.0
+
+    def merge(self, other: DomainAccumulator) -> None:
+        self.sources |= other.sources
+        self.evidence.extend(other.evidence)
+        self.cert_org_names |= other.cert_org_names
+        o_first = _normalize_time(other.first_seen)
+        if o_first and (self.first_seen is None or o_first < self.first_seen):
+            self.first_seen = o_first
+        o_last = _normalize_time(other.last_seen)
+        if o_last and (self.last_seen is None or o_last > self.last_seen):
+            self.last_seen = o_last
+        self.resolves = self.resolves or other.resolves
+        if self.rdap_org is None and other.rdap_org is not None:
+            self.rdap_org = other.rdap_org
+
+    def update_times(self, not_before: object, not_after: object) -> None:
+        nb = _normalize_time(not_before)
+        na = _normalize_time(not_after)
+        if nb and (self.first_seen is None or nb < self.first_seen):
+            self.first_seen = nb
+        if na and (self.last_seen is None or na > self.last_seen):
+            self.last_seen = na
